Remove support for go 1.8
go 1.8 reached EoL in February 2018.
